megerősítésekhez modal (mentésnél, kijelentkezésnél, törlésnél, ...); konzultációs csoportok felvétele (belső konzulens); konzultációs alkalmak felvétele, szerkesztése (belső konzulens); jquery ui datepicker helyett bootstrap datepicker; első diplomakurzus teljesítésének rögzítése nem külön táblában, hanem a témával egybe; első diplomakurzus teljesítésének rögzítése belső konzulens

+++ src/Controller/InternalConsultant/ConsultationsController.php
@@ -33,44 +33,35 @@ class ConsultationsController extends AppController
         
         $consultations = $this->Consultations->find('all', ['conditions' => ['thesis_topic_id' => $thesisTopic->id], 'order' => ['created' => 'DESC']]);
         
-        $this->set(compact('consultations'));
+        $this->set(compact('consultations', 'thesisTopic'));
     }
-
-    /**
-     * View method
-     *
-     * @param string|null $id Consultation id.
-     * @return \Cake\Http\Response|void
-     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
-     */
-    public function view($id = null)
-    {
-        $consultation = $this->Consultations->get($id, [
-            'contain' => ['Theses', 'ConsultationOccasions']
-        ]);
-
-        $this->set('consultation', $consultation);
-    }
-
+    
     /**
      * Add method
      *
      * @return \Cake\Http\Response|null Redirects on successful add, renders view otherwise.
      */
-    public function add()
+    public function add($thesis_topic_id = null)
     {
-        $consultation = $this->Consultations->newEntity();
-        if ($this->request->is('post')) {
-            $consultation = $this->Consultations->patchEntity($consultation, $this->request->getData());
-            if ($this->Consultations->save($consultation)) {
-                $this->Flash->success(__('The consultation has been saved.'));
-
-                return $this->redirect(['action' => 'index']);
-            }
-            $this->Flash->error(__('The consultation could not be saved. Please, try again.'));
+        $this->loadModel('Users');
+        $user = $this->Users->get($this->Auth->user('id'), ['contain' => ['InternalConsultants']]);
+        
+        $thesisTopic = $this->Consultations->ThesisTopics->find('all', ['conditions' => ['ThesisTopics.id' => $thesis_topic_id, 'internal_consultant_id' => $user->has('internal_consultant') ? $user->internal_consultant->id : '',
+                                                                          'thesis_topic_status_id' => 8], //Belső konzulenshez tartozik és elfogadott
+                                                                          ])->first();
+        
+        if(empty($thesisTopic)){
+            $this->Flash->error(__('A téma, amelyhez konzultációs csoportot szeretne adni, nem elérhető. Vagy nem létezik a téma, vagy nem Ön a belső konzulense.'));
+            return $this->redirect(['controller' => 'ThesisTopics', 'action' => 'index']);
         }
-        $theses = $this->Consultations->Theses->find('list', ['limit' => 200]);
-        $this->set(compact('consultation', 'theses'));
+        
+        $consultation = $this->Consultations->newEntity();
+        $consultation->thesis_topic_id = $thesisTopic->id;
+        
+        if ($this->Consultations->save($consultation)) $this->Flash->success(__('Mentés sikeres.'));
+        else $this->Flash->error(__('Mentés sikertelen. Próbálja újra!'));
+
+        return $this->redirect(['action' => 'index', $thesisTopic->id]);
     }
 
     /**
@@ -108,13 +99,31 @@ class ConsultationsController extends AppController
     public function delete($id = null)
     {
         $this->request->allowMethod(['post', 'delete']);
-        $consultation = $this->Consultations->get($id);
+        $consultation = $this->Consultations->find('all', ['conditions' => ['id' => $id]])->first();
+        
+        if(empty($consultation)){
+            $this->Flash->error(__('Konzultációs csoport nem létezik.'));
+            return $this->redirect(['controller' => 'ThesisTopics', 'action' => 'index']);
+        }
+        
+        $this->loadModel('Users');
+        $user = $this->Users->get($this->Auth->user('id'), ['contain' => ['InternalConsultants']]);
+        
+        $thesisTopic = $this->Consultations->ThesisTopics->find('all', ['conditions' => ['ThesisTopics.id' => $consultation->thesis_topic_id, 'internal_consultant_id' => $user->has('internal_consultant') ? $user->internal_consultant->id : '',
+                                                                          'thesis_topic_status_id' => 8], //Belső konzulenshez tartozik és elfogadott
+                                                                          ])->first();
+        
+        if(empty($thesisTopic)){
+            $this->Flash->error(__('A téma, amelynél konzultációs csoportot szeretne törölni, nem elérhető. Vagy nem létezik a téma, vagy nem Ön a belső konzulense.'));
+            return $this->redirect(['controller' => 'ThesisTopics', 'action' => 'index']);
+        }
+        
         if ($this->Consultations->delete($consultation)) {
-            $this->Flash->success(__('The consultation has been deleted.'));
+            $this->Flash->success(__('Törlés sikeres.'));
         } else {
-            $this->Flash->error(__('The consultation could not be deleted. Please, try again.'));
+            $this->Flash->error(__('Törlés sikertelen. Próbálja újra!'));
         }
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect(['action' => 'index', $consultation->thesis_topic_id]);
     }
 }
